refactor: get요청 query 함수 커스텀 훅 분리
