+++ src/components/Room/Question/QuestionList.jsx
@@ -1,10 +1,8 @@
 import { styled } from 'styled-components';
-import axios from 'axios';
-import { useQuery } from '@tanstack/react-query';
 
-// import { useRoomDataStore } from '../../../stores/Room/useRoomStore';
 import { useParams } from 'react-router-dom';
 import QuestionListItem from './QuestionListItem';
+import useQuestionsQuery from '../../../hooks/useQuestionsQuery';
 
 const QuestionListContainer = styled.ul`
   width: 100%;
@@ -12,18 +10,9 @@ const QuestionListContainer = styled.ul`
   gap: 40px;
 `;
 
-const getQuestionList = async (roomId) => {
-  const list = await axios.get(`${process.env.REACT_APP_API_URL}/rooms/${roomId}/questions`);
-
-  return list;
-};
 function QuestionList() {
   const { roomId } = useParams();
-  const { data, isSuccess, isError } = useQuery({
-    queryKey: [`question-${roomId}`],
-    queryFn: () => getQuestionList(roomId),
-    staleTime: 20000,
-  });
+  const { data, isSuccess, isError } = useQuestionsQuery(roomId);
 
   if (isError) console.log('QuestionList error');
   if (isSuccess) {
